requirements.txt
Django==1.5.4
django-filter==0.7
django-uuidfield==0.4.0
djangorestframework==2.3.8
python-dateutil==1.5
